papershaper/lamda/requirements.txt
pandas
filelock
google-generativeai